+++ aurora/api/au_api.hpp
@@ -6,4 +6,5 @@
 namespace aurora {
     void api_register_imgui(lua_State *L);
     void api_register_filesystem(lua_State *L);
+    void api_register_glm(lua_State *L);
 }